Implemented Defs.go
Place all new constants in Defs.go. When you’d like to use it, import
"github.com/arminm/multegula/defs" and use refs.CONSTANT to use CONSTANT
